+++ CHANGELOG.md
@@ -1,9 +1,16 @@
-## 0.0.1
+## 0.0.3
 
-- Initial version.
+- **Improvement**: Added input validation for Pix Key, Merchant Name, City, Amount, and TxId.
+- **New Feature**: Added support for custom `txId` (Transaction ID).
+- **Cleanup**: Removed unused dependencies (`crc`, `slugify`, `luthor`).
+- **Documentation**: Updated README and added improved example.
 
 ## 0.0.2
 
 - Added  `pointOfInitiationMethod`
 - Added `postalCode`
 - Remove upper casing in PixKey
+
+## 0.0.1
+
+- Initial version.